fix(templates/distribution/policy): don't patch kyverno config when type is gatekeeper
Don't patch Kyverno's config for excluding namespaces when the policy
type is Gatekeeper.

Previously, if the furyctl.yaml included the configuration for excluding
namespaces in Kyverno, the patch would be generated even if the
policy.type was Gatekeeper.
Now, when the policy type is Gatekeeper, the patch to exclude namespaces
in Kyverno won't be generated even if the configuration parameter is
present in the furyctl.yaml file.

Other changes included:
- Simplified the patching logic of the enforcementAction for
Gatekeeper's constratains using the group instead of patching each one
individually, reducing the risk of forgetting to patch some constraint
if a new one is added or not patching a new constraint that could be
added.
- Move everything that is possible under the `patches` key in the
kustomizatino file, being that patchesStrategicMerge has been deprecated
in newer versions of kustomize. Notice that due to a bug the patch for
the infra nodes selection cannot be moved yet.