fix(migrations/policy): delete objects before crds
When switching from kyverno to none, delete first the objects using the
APIs of the CRDs and then the CRDs. Otherwise we could sometimes end up
in a race condition where the objects cannot be deleted because the APIs
are not available anymore.

Fixes https://github.com/sighupio/fury-distribution/issues/335